+++ src/locales/i18n/base/index.js
@@ -1,11 +1,11 @@
 // This file is the base for translation
 
 import components from './components'
-import pages from './pages'
+import views from './views'
 import errors from './errors'
 
 export default {
   components,
-  pages,
+  views,
   errors
 }