+++ src/locales/i18n/en_US/index.js
@@ -1,9 +1,9 @@
 import components from './components'
-import pages from './pages'
+import views from './views'
 import errors from './errors'
 
 export default {
   components,
-  pages,
+  views,
   errors
 }