+++ src/views/Preferences/Greetings.vue
@@ -1,9 +1,9 @@
 <template>
   <v-ons-page id="greetings">
     <div class="content">
-      <h1 style="text-align:center">{{ $t('lang.pages.greetings.greetings') }}</h1>
+      <h1 style="text-align:center">{{ $t('lang.views.greetings.greetings') }}</h1>
       <br />
-      <h3>{{ $t('lang.pages.greetings.main') }}</h3>
+      <h3>{{ $t('lang.views.greetings.main') }}</h3>
       <br />
       <div class="homePageButton">
         <v-ons-button
@@ -13,7 +13,7 @@
           :disabled="false"
           ripple="true"
           @click.prevent="toHomePage"
-        >{{ $t('lang.pages.greetings.button') }}</v-ons-button>
+        >{{ $t('lang.views.greetings.button') }}</v-ons-button>
       </div>
     </div>
   </v-ons-page>
@@ -22,7 +22,7 @@
 <script>
 import { mapGetters, mapMutations } from 'vuex'
 import { updateDoc } from '@services/database'
-import AppSplitter from '@pages/AppSplitter'
+import AppSplitter from '@views/AppSplitter'
 
 export default {
   name: 'greetings',